pyproject.toml
[project]
name = "llm-api-daily-benchmark"
version = "0.1.0"
description = "LLM API daily benchmark - track model performance over time"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "streamlit>=1.28",
    "datasets>=2.14",
    "pyyaml>=6.0",
    "requests>=2.31",
    "httpx>=0.27",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "matplotlib>=3.7",
    "scipy>=1.11",
    "jinja2>=3.1",
    "sympy>=1.12",
]

[project.scripts]
benchmark = "benchmark.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["benchmark"]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
# BigCodeBench-Hard 轻量题所需的第三方执行依赖
benchmark = [
    "seaborn>=0.12",
    "beautifulsoup4>=4.12",
    "psutil>=5.9",
    "flask>=3.0",
    "flask-login>=0.6",
    "flask-wtf>=1.2",
    "flask-mail>=0.9",
    "wtforms>=3.1",
    "werkzeug>=3.0",
    "statsmodels>=0.14",
    "pytz>=2023.3",
    "lxml>=4.9",
    "openpyxl>=3.1",
    "xlwt>=1.3",
    "python-docx>=1.1",
    "pyquery>=2.0",
    "regex>=2023.6",
    "chardet>=5.2",
    "Pillow>=10.0",
    "rsa>=4.9",
]